+++ PKHeX.Core/PKM/PK4.cs
@@ -7,12 +7,12 @@ namespace PKHeX.Core
     /// <summary> Generation 4 <see cref="PKM"/> format. </summary>
     public sealed class PK4 : G4PKM
     {
-        private static readonly byte[] Unused =
+        private static readonly ushort[] Unused =
         {
             0x42, 0x43, 0x5E, 0x63, 0x64, 0x65, 0x66, 0x67, 0x87
         };
 
-        public override IReadOnlyList<byte> ExtraBytes => Unused;
+        public override IReadOnlyList<ushort> ExtraBytes => Unused;
 
         public override int SIZE_PARTY => PKX.SIZE_4PARTY;
         public override int SIZE_STORED => PKX.SIZE_4STORED;
